Add script to print group IPs
The group assignment demo needs the group VM IPs for the
http-load-generator, and for the prometheus scraper. This commit adds a
new `print_group_ips.sh` script that can be used to obtain the VM
IPs, and modifies the existing `print_client_config` to only output a
VMs IP given its identifier.